# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MyDay Kanban API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Column {
  id: number;
  title: string;
  position: number;
  createdAt: string;
}

export interface CreateColumnInput {
  title: string;
  position: number;
}

export interface UpdateColumnInput {
  title?: string;
  position?: number;
}

export interface Card {
  id: number;
  columnId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  position: number;
  /** @nullable */
  priority?: string | null;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
}

export interface CreateCardInput {
  columnId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  position: number;
  /** @nullable */
  priority?: string | null;
  /** @nullable */
  dueDate?: string | null;
}

export interface UpdateCardInput {
  columnId?: number;
  title?: string;
  /** @nullable */
  description?: string | null;
  position?: number;
  /** @nullable */
  priority?: string | null;
  /** @nullable */
  dueDate?: string | null;
}

export interface DeleteResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}