# lib/api-zod/src/generated/types/card.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MyDay Kanban API
 * OpenAPI spec version: 0.1.0
 */

export interface Card {
  id: number;
  columnId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  position: number;
  /** @nullable */
  priority?: string | null;
  /** @nullable */
  dueDate?: string | null;
  createdAt: Date;
}