# lib/api-zod/src/generated/types/column.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MyDay Kanban API
 * OpenAPI spec version: 0.1.0
 */

export interface Column {
  id: number;
  title: string;
  position: number;
  createdAt: Date;
}