# lib/api-zod/src/generated/types/createColumnInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MyDay Kanban API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateColumnInput {
  title: string;
  position: number;
}