# lib/api-zod/src/generated/types/updateColumnInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MyDay Kanban API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateColumnInput {
  title?: string;
  position?: number;
}